requirements.txt
fastapi
uvicorn
numpy
opencv-python-headless
Pillow
googletrans==4.0.0-rc1